Generated needed classes, interfaces, custom pipes, services, subcomponents, and a summary dash panel component that displays an initial stock portfolio.  The services get current stock prices for the symbols in the portfolio and the application then caclulates and displays the total value of each stock.  The summary also shows the desired portfolio and all associated information.
ToDo: will build the components to rebalance the portfolio and update the summary panel. Part of this component will display which stocks have been bought and sold in their respective quantities, and the worth of those transactions. Planning to add charts to help show the break down of the portfolio, and visualize the total worth of the different portfolios.

Commits from herein will now give detailed breakdowns of each file's associated changes.